+++ app/src/main/java/woowacourse/paint/presentation/ui/canvas/painttool/stroking/Line.kt
@@ -1,8 +1,9 @@
-package woowacourse.paint.presentation.ui.canvas.painttool
+package woowacourse.paint.presentation.ui.canvas.painttool.stroking
 
 import woowacourse.paint.presentation.ui.canvas.Palette
+import woowacourse.paint.presentation.ui.canvas.painttool.PaintTool
 
-class Line(palette: Palette) : StrokePaintTool(palette) {
+class Line(palette: Palette) : StrokingPaintTool(palette) {
 
     override fun nextPath(): PaintTool = Line(palette)
