+++ app/src/main/java/woowacourse/paint/presentation/ui/canvas/painttool/stroking/StrokingPaintTool.kt
@@ -1,11 +1,12 @@
-package woowacourse.paint.presentation.ui.canvas.painttool
+package woowacourse.paint.presentation.ui.canvas.painttool.stroking
 
 import android.graphics.Canvas
 import android.graphics.Paint
 import android.graphics.Path
 import woowacourse.paint.presentation.ui.canvas.Palette
+import woowacourse.paint.presentation.ui.canvas.painttool.PaintTool
 
-abstract class StrokePaintTool(
+abstract class StrokingPaintTool(
     final override val palette: Palette = Palette(),
 ) : PaintTool {
